IRIS-3881 | moved SingInPage and RegisterPage to auth page objects package
