+++ src/test/java/com/wikia/webdriver/pageobjectsfactory/pageobject/auth/RegisterPage.java
@@ -1,4 +1,4 @@
-package com.wikia.webdriver.elements.mercury.pages.login;
+package com.wikia.webdriver.pageobjectsfactory.pageobject.auth;
 
 import com.wikia.webdriver.common.core.elemnt.Wait;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.WikiBasePageObject;
